+++ Wastory/Wastory/ViewModel/ArticleViewModel/ArticleViewModel.swift
@@ -77,7 +77,7 @@ import RichTextKit
         do {
             let response = try await NetworkRepository.shared.getDraft(draftID: draftID)
             title = response.title
-            if let loadedText = DataTotext(response.content) {
+            if let loadedText = RichTextHandler.DataTotext(response.content) {
                 let restoredText = await RichTextImageHandler.restoreImage(loadedText)
                 text = restoredText
             }
@@ -93,7 +93,7 @@ import RichTextKit
     
     func storeDraft() async {
         let processedText = await RichTextImageHandler.convertImage(text)
-        if let dataText = textToData(processedText) {
+        if let dataText = RichTextHandler.textToData(processedText) {
             if currentDraftID < 0 {
                 do {
                     let response = try await NetworkRepository.shared.postDraft(title: title, content: dataText)
@@ -113,93 +113,5 @@ import RichTextKit
     
     func toggleIsDraftSheetPresent() {
         isDraftSheetPresent.toggle()
-    }
-    
-    // MARK: - Converters
-    // MARK: Main Converter
-    func textToData(_ text: NSAttributedString) -> String? {
-        do {
-            let data = try NSKeyedArchiver.archivedData(withRootObject: text, requiringSecureCoding: false)
-            return data.base64EncodedString()
-        } catch {
-            print("Failed to archive NSAttributedString: \(error)")
-            return nil
-        }
-    }
-    
-    func DataTotext(_ data: String) -> NSAttributedString? {
-        if let text = Data(base64Encoded: data) {
-            do {
-                return try NSAttributedString(data: text, format: .archivedData)
-            } catch {
-                print("Failed to load NSAttributedString: \(error)")
-            }
-        }
-        return nil
-    }
-    
-    // MARK: Sub Converter (may not use)
-    func HTMLTotext(_ htmlString: String) -> NSAttributedString? {
-        guard let htmlData = htmlString.data(using: .utf8) else { return nil }
-        do {
-            let attributedString = try NSAttributedString(
-                data: htmlData,
-                options: [
-                    .documentType: NSAttributedString.DocumentType.html,
-                    .characterEncoding: String.Encoding.utf8.rawValue
-                ],
-                documentAttributes: nil
-            )
-            return attributedString
-        } catch {
-            print("Error converting HTML to NSAttributedString: \(error)")
-        }
-        return nil
-    }
-    
-    func textToHTML(_ text: NSAttributedString) -> String? {
-        do {
-            let htmlData = try text.data(
-                from: NSRange(location: 0, length: text.length),
-                documentAttributes: [
-                    .documentType: NSAttributedString.DocumentType.html,
-                    .characterEncoding: String.Encoding.utf8.rawValue
-                ]
-            )
-            if let htmlString = String(data: htmlData, encoding: .utf8) {
-                return htmlString
-            }
-        }
-        catch {
-            print("Error convertnig Rich Text to HTML: \(error)")
-        }
-        return nil
-    }
-    
-    func textToRTF(_ text: NSAttributedString) -> String? {
-        do {
-            let rtfData = try text.data(
-                from: NSRange(location: 0, length: text.length),
-                documentAttributes: [.documentType: NSAttributedString.DocumentType.rtf])
-            return rtfData.base64EncodedString()
-        }
-        catch {
-            print("Error convertnig Rich Text to RTF: \(error)")
-        }
-        return nil
-    }
-    
-    func RTFTotext(_ rtfString: String) -> NSAttributedString? {
-        do {
-            let rtfData = Data(base64Encoded: rtfString)
-            return try NSAttributedString(
-                data: rtfData!,
-                options: [.documentType: NSAttributedString.DocumentType.rtf],
-                documentAttributes: nil)
-        }
-        catch {
-            print("Error convertnig RTF to Rich Text: \(error)")
-        }
-        return nil
     }
 }
